Reduce the slopegraph height to 1200.
Now that the slopegraph axes are consistent, we don't need as large of a graph. There still seems to be an issue
if the height is less than 1200, e.g. some labels will show an increasing slope even if the percentage decreases.
This is probably because the labels are too tightly clustered together.